+++ libsolidity/codegen/ArrayUtils.h
@@ -74,7 +74,7 @@ public:
 	void popStorageArrayElement(ArrayType const& _type) const;
 	/// Appends a loop that clears a sequence of storage slots of the given type.
 	/// Stack pre: start_ref slot_count
-	/// Stack post: start_ref
+	/// Stack post:
 	void clearStorageLoop(Type const* _type) const;
 	/// Converts length to size (number of storage slots or calldata/memory bytes).
 	/// if @a _pad then add padding to multiples of 32 bytes for calldata/memory.