+++ libyul/backends/evm/ControlFlowGraphBuilder.h
@@ -23,6 +23,8 @@
 #include <libyul/backends/evm/ControlFlowGraph.h>
 #include <libyul/ControlFlowSideEffects.h>
 
+#include <unordered_map>
+
 namespace solidity::yul
 {
 
@@ -56,7 +58,7 @@ private:
 	ControlFlowGraphBuilder(
 		CFG& _graph,
 		AsmAnalysisInfo const& _analysisInfo,
-		std::map<FunctionDefinition const*, ControlFlowSideEffects> const& _functionSideEffects,
+		std::unordered_map<FunctionDefinition const*, ControlFlowSideEffects> const& _functionSideEffects,
 		Dialect const& _dialect
 	);
 	void registerFunction(FunctionDefinition const& _function);
@@ -79,7 +81,7 @@ private:
 	);
 	CFG& m_graph;
 	AsmAnalysisInfo const& m_info;
-	std::map<FunctionDefinition const*, ControlFlowSideEffects> const& m_functionSideEffects;
+	std::unordered_map<FunctionDefinition const*, ControlFlowSideEffects> const& m_functionSideEffects;
 	Dialect const& m_dialect;
 	CFG::BasicBlock* m_currentBlock = nullptr;
 	Scope* m_scope = nullptr;