+++ test/libsolidity/semanticTests/array/copying/array_copy_including_array.sol
@@ -36,11 +36,11 @@ contract c {
 // ----
 // test() -> 0x02000202
 // gas irOptimized: 4560468
-// gas legacy: 4536566
-// gas legacyOptimized: 4456759
+// gas legacy: 4536539
+// gas legacyOptimized: 4456732
 // storageEmpty -> 1
 // clear() -> 0, 0
 // gas irOptimized: 4488719
-// gas legacy: 4407780
-// gas legacyOptimized: 4385089
+// gas legacy: 4407759
+// gas legacyOptimized: 4385068
 // storageEmpty -> 1