+++ test/libsolidity/semanticTests/array/pop/array_pop_array_transition.sol
@@ -24,6 +24,6 @@ contract c {
 // ----
 // test() -> 1, 2, 3
 // gas irOptimized: 1828383
-// gas legacy: 1841995
-// gas legacyOptimized: 1821687
+// gas legacy: 1841877
+// gas legacyOptimized: 1821569
 // storageEmpty -> 1