+++ Changelog.md
@@ -1,11 +1,21 @@
-### 0.8.34 (unreleased)
+### 0.8.35 (unreleased)
 
 Language Features:
 
 Compiler Features:
-* Yul Optimizer: Remove redundant prerequisite steps from the default optimizer sequence.
 
 Bugfixes:
+
+
+### 0.8.34 (2026-02-18)
+
+Important Bugfixes:
+* Yul IR Code Generation: Fix a bug that could result in clearing a storage variable instead of a transient storage variable at the same position in the layout (and vice-versa).
+
+
+Compiler Features:
+* Yul Optimizer: Improve performance of control flow side effects collector and function references resolver.
+* Yul Optimizer: Remove redundant prerequisite steps from the default optimizer sequence.
 
 
 ### 0.8.33 (2025-12-18)